requirements-sunny.txt
numpy
sounddevice
faster-whisper
soundfile
ollama
elevenlabs
websocket-client
pyyaml
